Complete DSM text styles
