+++ src/_design-tokens-temp.scss
@@ -1,0 +1,124 @@
+/*
+    Hardcoded design tokens placed here until new invision
+    DSM provides design token functionality.
+*/
+
+
+/*
+    Colours
+*/
+
+$colour-blue-deep-80: #4b426d !default;
+
+
+/*
+    Text Styles
+*/
+
+@mixin font-title-1 () {
+    font-size: 24px;
+    line-height: 30px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 600;
+}
+
+@mixin font-title-2 () {
+    font-size: 28px;
+    line-height: 32px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 600;
+}
+
+@mixin font-title-3 () {
+    font-size: 36px;
+    line-height: 40px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 600;
+}
+
+@mixin font-title-4 () {
+    font-size: 48px;
+    line-height: 60px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 600;
+}
+
+@mixin font-title-5 () {
+    font-size: 60px;
+    line-height: 60px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 600;
+}
+
+@mixin font-body-1 () {
+    font-size: 12px;
+    line-height: 15px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 400;
+}
+
+@mixin font-body-1-bold () {
+    font-size: 12px;
+    line-height: 15px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 600;
+}
+
+@mixin font-body-2 () {
+    font-size: 16px;
+    line-height: 20px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 400;
+}
+
+@mixin font-body-2-bold () {
+    font-size: 16px;
+    line-height: 20px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 600;
+}
+
+@mixin font-body-3 () {
+    font-size: 20px;
+    line-height: 30px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 400;
+}
+
+@mixin font-body-3-bold () {
+    font-size: 20px;
+    line-height: 30px;
+    color: $colour-blue-deep-80;
+    font-family: 'FS Lucas';
+    font-style: normal;
+    font-weight: 600;
+}
+
+@mixin quote-text () {
+    font-size: 24px;
+    line-height: 40px;
+    color: $colour-blue-deep-80;
+    font-family: Georgia;
+    font-style: italic;
+    font-weight: 400;
+}
