--- stories/heading.stories.js
@@ -1,24 +1,0 @@
-import React from "react";
-
-export default { title: "Typography" };
-
-export const Headings = () => (
-  <div className="m-s">
-    <h1 className="h-xl">.h-xl</h1>
-    <h1 className="h-l">.h-l</h1>
-    <h1 className="h-m">.h-m</h1>
-  </div>
-);
-
-export const Body = () => (
-  <div className="m-s">
-    <h1 className="body-xl">.body-xl</h1>
-    <h1 className="body-l">.body-l</h1>
-    <h1 className="body-m">.body-m</h1>
-    <h1 className="body-s">.body-s</h1>
-    <h1 className="body-xl-b">.body-xl-b</h1>
-    <h1 className="body-l-b">.body-l-b</h1>
-    <h1 className="body-m-b">.body-m-b</h1>
-    <h1 className="body-s-b">.body-s-b</h1>
-  </div>
-);
